refactor: creating shared types ts file for HTMLManager and GameManager.
