+++ public/ts/types/shared.types.ts
@@ -1,0 +1,6 @@
+export type Board = number[][]
+
+export type SlotCoords = {
+    x: number,
+    y: number
+}
